implement bounding for heatmap
