+++ src/components/data-exploration/CellInfo.jsx
@@ -1,24 +1,9 @@
 import React, { useCallback, useState } from 'react';
 import { Card } from 'antd';
 import PropTypes from 'prop-types';
-
-const EM = 16; // px
-const CELLINFO_Y_PADDING = 2 * EM;
+import getCellInfoCoordinates from 'utils/data-exploration/getCellInfoCoordinates';
 
 const cellInfoStyle = { fontSize: '0.75rem' };
-
-const getCellinfoCoordinates = (coordinates, el, boundingX, boundingY) => {
-  const popupWidth = el?.offsetWidth;
-  const popupHeight = el?.offsetHeight;
-
-  const invertX = () => coordinates.x + popupWidth > boundingX;
-  const invertY = () => coordinates.y + popupHeight + CELLINFO_Y_PADDING > boundingY;
-
-  const left = invertX() ? coordinates.x - (popupWidth + EM) : coordinates.x + EM;
-  const top = invertY() ? coordinates.y - (popupHeight + EM) : coordinates.y + EM;
-
-  return { left, top };
-};
 
 const CellInfo = (props) => {
   const {
@@ -31,7 +16,7 @@ const CellInfo = (props) => {
     if (el) setTooltipEl(el?.firstChild);
   }, []);
 
-  const { left, top } = getCellinfoCoordinates(
+  const { left, top } = getCellInfoCoordinates(
     coordinates,
     tooltipEl,
     containerWidth,
